Cars API improvements
