+++ src/DataFixtures/UserFixtures.php
@@ -57,12 +57,12 @@ final class UserFixtures extends Fixture implements DependentFixtureInterface
         return [
             // data = [$firstName, $lastName, $password, $verified, $is_account, $phone, $email, $roles]
             ['Admin', 'Admin', 'admin', true, true, '(123)555-1111', 'admin@admin.com', ['ROLE_ADMIN', 'ROLE_USER']],
-            ['Dealer2', 'Dealer2', 'test', true, true, '(456)555-2222', 'dealer2@test.com', ['ROLE_USER']],
-            ['Dealer3', 'Dealer3', 'test', true, true, '(456)555-3333', 'dealer3@test.com', ['ROLE_USER']],
-            ['Staff4', 'Staff4', 'test', true, true, '(456)555-4444', 'staff4@test.com', ['ROLE_USER']],
-            ['Staff5', 'Staff5', 'test', true, true, '(456)555-5555', 'staff5@test.com', ['ROLE_USER']],
-            ['Buyer5', 'Buyer5', 'test', true, true, '(456)555-6666', 'buyer5@test.com', ['ROLE_USER']],
-            ['Buyer6', 'Buyer6', 'test', true, false, '(456)555-7777', 'buyer6@test.com', ['ROLE_USER']],
+            ['Dealer2', 'Dealer2', 'test', true, true, '(456)555-2222', 'dealer2@test.com', ['ROLE_DEALER']],
+            ['Dealer3', 'Dealer3', 'test', true, true, '(456)555-3333', 'dealer3@test.com', ['ROLE_DEALER']],
+            ['Staff4', 'Staff4', 'test', true, true, '(456)555-4444', 'staff4@test.com', ['ROLE_STAFF']],
+            ['Staff5', 'Staff5', 'test', true, true, '(456)555-5555', 'staff5@test.com', ['ROLE_STAFF']],
+            ['Buyer5', 'Buyer5', 'test', true, true, '(456)555-6666', 'buyer5@test.com', ['ROLE_BUYER']],
+            ['Buyer6', 'Buyer6', 'test', true, false, '(456)555-7777', 'buyer6@test.com', ['ROLE_BUYER']],
         ];
     }
