+++ src/Entity/Trait/CreatedAtTrait.php
@@ -11,9 +11,6 @@ trait CreatedAtTrait
 {
     #[ORM\Column(type: 'datetime_immutable', options: ['default' => 'CURRENT_TIMESTAMP'])]
     private ?DateTimeImmutable $created_at;
-    
-    #[ORM\Column(type: 'datetime_immutable', options: ['default' => 'CURRENT_TIMESTAMP'])]
-    private ?DateTimeImmutable $modified_at;
 
     public function getCreatedAt(): ?DateTimeImmutable
     {
@@ -23,18 +20,6 @@ trait CreatedAtTrait
     public function setCreatedAt(DateTimeImmutable $created_at): self
     {
         $this->created_at = $created_at;
-
-        return $this;
-    }
-
-    public function getModifiedAt(): ?DateTimeImmutable
-    {
-        return $this->modified_at;
-    }
-
-    public function setModifiedAt(DateTimeImmutable $modified_at): self
-    {
-        $this->modified_at = $modified_at;
 
         return $this;
     }
