+++ src/Form/Type/ProductCustomType.php
@@ -5,7 +5,8 @@ declare(strict_types=1);
 namespace App\Form\Type;
 
 use App\Entity\Category;
-use App\Entity\Product;
+// use App\Entity\Product;
+use App\Entity\ProductDetails;
 use App\Repository\CategoryRepository;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
@@ -28,53 +29,7 @@ class ProductCustomType extends AbstractType
                 'mapped' => true,
                 'required' => false,
             ])
-            ->add('title', TextType::class, [
-                'label' => 'label.title',
-                'attr' => [
-                    'placeholder' => 'label.title',
-                    'class' => 'form-control',
-                ],
-                'mapped' => true,
-                'required' => true,
-            ])
-            ->add('description', TextType::class, [
-                'label' => 'label.description',
-                'attr' => [
-                    'placeholder' => 'label.description',
-                    'class' => 'form-control',
-                ],
-                'mapped' => true,
-                'required' => false,
-            ])
-            ->add('price', TextType::class, [
-                'label' => 'label.price',
-                'attr' => [
-                    'placeholder' => 'label.price',
-                    'class' => 'form-control',
-                ],
-                'mapped' => true,
-                'required' => false,
-            ])
-            ->add('category', EntityType::class, [
-                'class' => Category::class,
-                'placeholder' => 'label.category',
-                'choice_label' => 'name',
-                'label' => 'label.category',
-                'mapped' => true,
-                'required' => false,
-                'group_by' => 'parent.name',
-                'query_builder' => function (CategoryRepository $cr) {
-                    return $cr->createQueryBuilder('c')
-                        ->where('c.parent IS NOT NULL')
-                        ->orderBy('c.name', 'ASC')
-                    ;
-                },
-            ])
-            ->add('images', FileType::class, [
-                'label' => false,
-                'mapped' => false,
-                'required' => false,
-            ])
+            
             ->add('suggested_vin', TextType::class, [
                 'label' => 'label.suggested_vin',
                 'attr' => [
@@ -344,11 +299,11 @@ class ProductCustomType extends AbstractType
             ->add('lane_centering_assistance', TextType::class, [ 'label' => 'label.make', 'attr' => [ 'placeholder' => 'label.make', 'class' => 'form-control', ], 'mapped' => true ])
         ;
 
-        $builder->get('title')->setRequired(true);
-        $builder->get('images')->setRequired(true);
-        $builder->get('category')->setRequired(true);
-        $builder->get('description')->setRequired(true);
-        $builder->get('price')->setRequired(true);
+        // $builder->get('title')->setRequired(true);
+        // $builder->get('images')->setRequired(true);
+        // $builder->get('category')->setRequired(true);
+        // $builder->get('description')->setRequired(true);
+        // $builder->get('price')->setRequired(true);
         $builder->get('suggested_vin')->setRequired(false);
         $builder->get('possible_values')->setRequired(false);
         $builder->get('vehicle_descriptor')->setRequired(false);
@@ -487,7 +442,7 @@ class ProductCustomType extends AbstractType
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            'data_class' => Product::class,
+            'data_class' => ProductDetails::class,
         ]);
     }
 }
