+++ templates/user/product/index.html.twig
@@ -34,7 +34,7 @@
 
                 {% for image in product.images | slice(0, 1) %}
 
-                    <img class="card-img-top" src="/uploads/images/small/{{ image.file }}"
+                    <img class="card-img-top" src="/uploads/images/{{ image.file }}"
                          alt="{{ product.title }}"
                          title="{{ product.title }}"/>
